Add ability to run external AP and switch modes, update readme  (#8)
* Create docker-compose-noap.yaml

* Update install.sh

* fix yes/no prompt

* Update install.sh

* don't kill network manager

* Update readme with AP changes

* Update README.md

* Update README.md

* Update README.md

* Update install.sh

* no color in question

* Create cron_docker_up for external AP startup

* do kill net manager for running AP on host

* Rename cron_docker_up to cron_dockerup

* Update install.sh

* consistency

* declare postgres version

* make the selection dialog prettier

* add postgres version

* refactor AP mode deployment add change ap script

* declare compose files to pull all images at install

* Update install.sh

* fix typo

* fix typo

* typo

* make executable

* fix order and networkmanager

* minor fixes

* fix network enable script

* add convenience script descriptions, clarify wifi

* fix branding

* fix branding for btt-recorder

* branding

+++ install.sh
@@ -20,6 +20,32 @@ if [[ $(grep -c "^$USER:" /etc/passwd) = 0 ]]; then
 	exit 1
 fi
 
+cd /home/$USER/btt-exchanger
+
+HEIGHT=10
+WIDTH=70
+CHOICE_HEIGHT=3
+TITLE="Access Point"
+MENU="Choose one of the following options:"
+
+OPTIONS=("1" "Run a Wifi Access Point on the server hardware"
+         "2" "Run an external access point (e.g. TP-Link TL-WR802N)")
+
+CHOICE=$(whiptail --title "$TITLE" \
+                --menu "$MENU" \
+                $HEIGHT $WIDTH $CHOICE_HEIGHT \
+                "${OPTIONS[@]}" \
+                3>&1 1>&2 2>&3)
+
+case $CHOICE in
+        1)
+            sudo cp /home/$USER/btt-exchanger/cron_dockerup_server_ap /etc/cron.d/cron_dockerup_server_ap
+            ;;
+        2)
+            sudo cp /home/$USER/btt-exchanger/cron_dockerup_external_ap /etc/cron.d/cron_dockerup_external_ap
+            ;;
+esac
+
 echo -e "${COLOR}----------| Installing neccessary software for the server... |----------${NC}"
 
 sudo apt update && sudo apt install -y curl python-pip apt-transport-https ca-certificates software-properties-common jq
@@ -27,13 +53,7 @@ curl -fsSL https://download.docker.com/linux/debian/gpg | sudo apt-key add -
 sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/debian $(lsb_release -cs) stable"
 sudo apt update && sudo apt install -y docker-ce
 sudo pip install docker-compose
-
-echo -e "${COLOR}----------| Inserting crontab task... |----------${NC}"
-
 sudo chmod -R +x /home/$USER/btt-exchanger/scripts
-sudo cp /home/$USER/btt-exchanger/cron_netsvc_and_dockerup /etc/cron.d/cron_netsvc_and_dockerup
-
-cd /home/$USER/btt-exchanger
 
 echo -e "${COLOR}----------| Downloading clients |----------${NC}"
 
@@ -43,10 +63,12 @@ rm -rf clients/
 unzip clients.zip
 
 echo -e "${COLOR}----------| Downloading TranslationRecorder|----------${NC}"
-URL=$(curl 'https://api.github.com/repos/wycliffeassociates/translationrecorder/releases?per_page=1' | jq -r '.[0] | .assets[].browser_download_url')
-curl -L $URL --output clients/translationRecorder.apk
+
+URL=$(curl 'https://api.github.com/repos/Bible-Translation-Tools/BTT-Recorder/releases?per_page=1' | jq -r '.[0] | .assets[].browser_download_url')
+curl -L $URL --output clients/bttRecorder.apk
 
 echo -e "${COLOR}----------| Downloading AdminTools into admintools dir |----------${NC}"
+
 rm -rf admintools/
 mkdir admintools
 cd admintools
@@ -64,8 +86,10 @@ cd /home/$USER/btt-exchanger
 sudo -- sh -c -e "echo '10.0.0.1	opentranslationtools.org' >> /etc/hosts"
 
 echo -e "${COLOR}----------| Pulling Docker Images... |----------${NC}"
-sudo docker-compose pull
 
-echo -e "${COLOR}**********| Installation complete. Rebooting... |**********${NC}"
-sleep 3
-sudo systemctl reboot
+sudo docker-compose -f docker-compose-ext-ap.yaml pull
+sudo docker-compose -f docker-compose-server-ap.yaml pull
+
+clear
+if [ $CHOICE == 1 ]; then read -n 1 -s -r -p "Installation complete. press Enter to reboot" && sleep 1 && sudo systemctl reboot; fi
+if [ $CHOICE == 2 ]; then read -n 1 -s -r -p "Installation complete. Please unplug network cable then press Enter to reboot" && sleep 1 && sudo systemctl reboot; fi
